==> Notebooks/00. baseline.ipynb <==
import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter

import pandas as pd
from tqdm.notebook import tqdm

data_path = '..\\Data\\In\\'
submissions_path =  '..\\Submissions\\'

## Чтение входных данных

Напишем простую функцию для чтения файла с частью вакансий:

def read_vacancies_part(part):
    with gzip.open(data_path+f'vacancies-{part:02}.json.gz', 'r') as fp:
        return json.loads(fp.read())

%%time

part = read_vacancies_part(1)
print(f'Всего вакансий в части 1: {len(part):,d}')

Пример содержимого вакансии:

part['2']

print(json.dumps(part['20'], indent=4, ensure_ascii=False))

Для этого бейзлана нам понадобятся только хэши работодателей, поэтому будет оставлять только их для экономии памяти:

%%time

vacancy_employers = {int(vacancy_id): vacancy['employer'] for vacancy_id, vacancy in part.items()}

for part_num in tqdm(range(2, 11)):
    part = read_vacancies_part(part_num)
    vacancy_employers.update({int(vacancy_id): vacancy['employer'] for vacancy_id, vacancy in part.items()})
    print(f'Всего вакансий в части {part_num}: {len(part):,d}')

print(f'Всего вакансий в датасете: {len(vacancy_employers):,d}')
del part

Прочитаем файл со специализациями из тренировочного множества:

train_specializations = pd.read_csv(data_path+'train_labels.csv.gz', compression='gzip')
train_specializations.head(5)

Для удобства превратим прочитанный датафрейм в словарь из id вакансии в список id специализаций:

train_specializations = {
    vacancy_id: list(map(int, specs[1:-1].split(',')))
    for vacancy_id, specs in train_specializations.set_index('vacancy_id')['specializations'].iteritems()
}

Наконец, прочитаем файл с id вакансий из тестового множества:

test_ids = pd.read_csv(data_path+'test_vacancy_ids.csv.gz', compression='gzip').values.ravel()

### Подготовка сабмита

Подготовим сабмит, основанный на интуиции о том, что работодатели часто создают вакансии с одними и теми же специализациями. Поэтому для каждой вакансии из тествого множества поставим в соответствие:
* список из top3 специализаций работодателя, если он встречался в тренировочном множесте
* одну самую популярную специализацию во всём тренировочном множестве, если мы не видели этого работодателя

Посчитаем встречаемость каждой специализации в целом по всем вакансиям, а также в отдельности по каждому работодателю:

%%time

common_counter = Counter()
employer_counters = defaultdict(lambda: Counter())

for vacancy_id, specs in tqdm(train_specializations.items()):
    employer_counter = employer_counters[vacancy_employers[vacancy_id]]
    for specialization in specs:
        common_counter[specialization] += 1
        employer_counter[specialization] += 1

Теперь нам известен top3 самых популярных специализаций по каждому работодателю, а также самая популярная специализация в датасете:

top_spec = common_counter.most_common(1)[0][0]
top3_specs_by_employer = {
    employer: list(map(itemgetter(0), counter.most_common(3)))
    for employer, counter in employer_counters.items()
}

print(f'top_spec: {top_spec}')
print('---')
for employer, employer_specs in islice(top3_specs_by_employer.items(), 10):
    print(f'{employer}: {employer_specs}')

Составим сабмит по описанному выше принципу и запишем его в таком же формате, как и файл со специализациями из тренировочного множества:

submissions_path

%%time
sample_submission = pd.DataFrame([
    (vacancy_id, top3_specs_by_employer.get(vacancy_employers[vacancy_id], [top_spec]))
    for vacancy_id in test_ids
], columns=['vacancy_id', 'specializations'])
sample_submission.to_csv(submissions_path+f'sample_submission_{"2020-05-25"}_{1}.csv.gz', index=False, compression='gzip')

Такой сабмит даёт **0.23546** mean f-score на public.
