==> Notebooks/02. Process Text.ipynb <==
%load_ext autoreload
%autoreload 2


import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

from workers import *

import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

data_path = '..\\Data\\In\\'
submissions_path =  '..\\Submissions\\'

#df.to_pickle("df.pkl")

df = pd.read_pickle("df.pkl")

df.head()

#df['html_tags'] = pqdm(df.description, extract_html, n_jobs=10)

df_train = df
del df

df_train.head()

# df_train = df[df.is_test == False].copy()
# del df

#df_train.to_parquet("df_tarin.pq")

# if __name__ ==  '__main__': 
#     df['html_tags'] = parallelize_on_rows(df.description, extract_html) 

# %%time
# df_train= pd.read_parquet("df_tarin.pq")

df_train.columns

%%time
if __name__ ==  '__main__':
    df_train['clear_text'] = parallelize_on_rows(df_train.description, remove_html) 

%%time
if __name__ ==  '__main__': 
    df_train['processed_text'] = parallelize_on_rows(df_train.clear_text, preprocess_text) 

qq('Precessing text done')

#test
df_train[df_train.vacancy_id == 2264555]['processed_text'].values

%%time
df_train.to_parquet("df_tarin.pq")

df_text = df_train[['vacancy_id', 'name', 'specializations', 'key_skills', 'employer', 'groups_of_spec', 'processed_text']].copy()

del df_train

#df_text = pd.read_parquet("df_text.pq")

df_text.head()

%%time
if __name__ ==  '__main__':
    df_text['name'] = parallelize_on_rows(df_text.name, preprocess_text)

df_text['key_skills'] = df_text['key_skills'].apply(' '.join)

%%time
if __name__ ==  '__main__':
    df_text['key_skills'] = parallelize_on_rows(df_text.key_skills, preprocess_text)

#from collections import Counter

#df_text.drop(columns=['text_list'], inplace = True)

%%time
df_text['key_list'] = df_text['key_skills'].apply(lambda x: x.split(' '))
df_text['name_list'] = df_text['name'].apply(lambda x: x.split(' '))
# df_text['text_list'] = df_text['processed_text'].apply(lambda x: x.split(' '))

df_text.head()

qq('all processing done')

%%time
df_text.to_parquet("df_text.pq")

%%time
df_text['text_list'] = df_text['processed_text'].apply(lambda x: x.split(' '))

import functools
import operator
full_text = functools.reduce(operator.iconcat, df_text['text_list'].to_list(), [])

word_count = Counter(full_text)

word_count

len(word_count)

word_count.most_common(60)

word_count.most_common()[:-100-1:-1]

import matplotlib.pyplot as plt

plt.hist([c for w, c in word_count.items() if c < 15000 and c > 10])

qq('Ready')
