==> Notebooks/03. Doc2Vec - gensim.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

qq('check')

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

#from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections

data_path = '..\\Data\\In\\'
submissions_path =  '..\\Submissions\\'

%%time
df = pd.read_parquet('train_text.pq')

%%time
df['text_list'] = df['processed_text'].apply(lambda x: x.split(' '))

#df['txt_words'] = df['text_list'].apply(len)

#df['txt_len'] = df['text_list'].apply(lambda x: sum([len(w) for w in x]))

df.head(2)

#df[df['txt_words'] > 300].shape

#df[df['txt_words'] < 200]['txt_words'].hist(bins=20)

def txt_2_td(x):
    x4 = [x[4]]
    x2 = list(x[2])
    x3 = (list(x[3]) if x[3][0] != '' else []) 
    doc = x4 + x2 + x3 + x[1]
    return TaggedDocument(doc, [str(x[0])])

def txt_2_td_n(x):
    x4 = [x[4]]
    x2 = list(x[2])
    x3 = (list(x[3]) if x[3][0] != '' else []) 
    doc = x4 + x2
    return TaggedDocument(doc, [str(x[0])])

# val = set(dill.load(open('validatin.bin','rb')))

# df['val']=df['vacancy_id'].isin(val)

# df[~df['val']].shape

%%time
document_text = df[['vacancy_id', 'text_list', 'name_list', 'key_list', 'employer']].apply(txt_2_td, axis = 1).to_list()

%%time
document_text_n = df[['vacancy_id', 'text_list', 'name_list', 'key_list', 'employer']].apply(txt_2_td_n, axis = 1).to_list()

qq('doc_text ready')

document_text_n[0]

len(document_text_n)

#from gensim.test.utils`a import get_tmpfile

#fname = get_tmpfile("my_doc2vec_model")

# Asserting model

import gc

gc.collect()

import model_doc2vec
from model_doc2vec import ProgressMeter

#document_text = dill.load(open('tagged_docs.bin','rb'))

model_doc2vec.train_model(document_text=document_text_n, vector_size=10, min_count=2, epochs=10, window = 5, seed = 42, workers = 8, negative = 10)
qq('done')

# 0.689 0.179 vector_size:250, min_count:2, epochs:10, window:5, seed:42, workers:12, negative:5
# 0.758 0.113 vector_size:300, min_count:2, epochs:10, window:5, seed:42, workers:6, negative:5
# 0.774 0.097 vector_size:300, min_count:2, epochs:10, window:7, seed:42, workers:4, negative:10
# 0.767 0.112 vector_size:300, min_count:2, epochs:10, window:10, seed:42, workers:4, negative:15
# 0.768 0.088 vector_size:300, min_count:2, epochs:10, window:10, seed:42, workers:4, negative:30
# 0.816 0.042 vector_size:300, min_count:2, epochs:10, window:7, seed:42, workers:4, negative:30
# 0.845 0.026 vector_size:300, min_count:2, epochs:10, window:7, seed:42, workers:4, negative:50

model_doc2vec.train_model(document_text=document_text, vector_size=400, min_count=2, epochs=20, window = 5, seed = 42, workers = 8, negative = 20)
qq('done')

model= Doc2Vec.load('models/vector_size:400_min_count:2_epochs:20_window:5_seed:42_workers:8_negative:20.bin')

model.docvecs.vectors_docs.shape

train_indexes = model.docvecs.index2entity

train_vectors = model.docvecs.vectors_docs

dill.dump(train_indexes, open('train_indexes.bin', 'wb'))

dill.dump(train_vectors, open('train_vectors.bin', 'wb'))

test_df = pd.read_parquet('test_text.pq')

%%time
test_df['text_list'] = test_df['processed_text'].apply(lambda x: x.split(' '))

def merge_txt(x):
    x4 = [x[4]]
    x2 = list(x[2])
    x3 = (list(x[3]) if x[3][0] != '' else []) 
    return x4 + x2 + x3 + x[1]

test_df['merged_text'] = test_df[['vacancy_id', 'text_list', 'name_list', 'key_list', 'employer']].apply(merge_txt, axis=1)

def inf_vec(txt):
    return model.infer_vector(txt)

import gc

gc.collect()

test_df.head()

test_vector = pqdm(test_df['merged_text'], inf_vec, n_jobs=11)

dill.dump(test_vector, open('test_vectors.bin', 'wb'))

dill.dump(test_df['vacancy_id'].to_list(), open('test_indexes.bin', 'wb'))

spec_df = df[['vacancy_id', 'specializations']].copy()

spec_df.set_index('vacancy_id', inplace=True)

dill.dump(spec_df.to_dict(orient='dict')['specializations'], open('train_spec.bin', 'wb'))

num_samples = 1000
TOPN = 10
doc_samples = random.sample(range(len(document_text)),num_samples)
ranks = []
for doc_id in doc_samples:
    inferred_vector = model.infer_vector(document_text[doc_id].words)
    sims = model.docvecs.most_similar([inferred_vector], topn=TOPN)
    try:
        ranks.append([docid for docid, sim in sims].index(document_text[doc_id].tags[0]))
    except:
        ranks.append(11)

# ranks = [x[1] for x in result]
# sims = [x[2] for x in result]

collections.Counter(ranks)

doc_id = random.randint(low_bound, low_bound+num_samples)
print('Document ({}): «{}»\n'.format(doc_id, ' '.join(document_text[doc_id].words)))
print(u'SIMILAR/DISSIMILAR DOCS PER MODEL %s:\n' % model)
for label, index in [('MOST', 0), ('SECOND-MOST', 1), ('MEDIAN', len(sims)//2), ('LEAST', len(sims) - 1)]:
    print(u'%s %s: «%s»\n' % (label, sims[index], ' '.join(document_text[sims[index][0]].words)))

# Pick a random document from the test corpus and infer a vector from the model
doc_id = random.randint(0, len(document_text) - 1)
inferred_vector = model.infer_vector(document_text[doc_id].words)
sims = model.docvecs.most_similar([inferred_vector], topn=len(model.docvecs))

docs_ids.index(sims[index][0])

len(sims)

tag_to_id = {v.tags[0]:i for i,v in enumerate(document_text)}

len(sims)-1

document_text[tag_to_id[sims[index][0]]]

tag_to_id

tag_to_id[sims[index][0]]

# Compare and print the most/median/least similar documents from the train corpus
print('Test Document ({}): «{}»\n'.format(doc_id, ' '.join(document_text[doc_id].words)))
print(u'SIMILAR/DISSIMILAR DOCS PER MODEL %s:\n' % model)
for label, index in [('MOST', 0), ('MEDIAN', len(sims)//2), ('LEAST', len(sims) - 1)]:
    print(u'%s %s: «%s»\n' % (label, sims[index], ' '.join(document_text[docs_ids.index(sims[index][0])].words)))
