==> Notebooks/04. Train_Test Split.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

qq('start')

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections

df = pd.read_parquet('df_tarin.pq', columns=['vacancy_id', 'employer'])

# df.columns
# Index(['vacancy_id', 'name', 'description', 'area_id', 'creation_date',
#        'work_schedule', 'work_experience', 'compensation_from',
#        'compensation_to', 'currency', 'key_skills', 'employer',
#        'specializations', 'is_test', 'part', 'employment', 'spec_len',
#        'groups_of_spec', 'len_groups', 'html_tags', 'clear_text',
#        'processed_text'],
#       dtype='object')

df[df.employer == '0000026a6f11']

df.shape

empl = df['employer'].value_counts().to_dict()

df['is_validation'] = False

df['num_samples'] = df['employer'].apply(lambda x: empl[x])

validation = []
valid_part = 0.2
for k,i in empl.items():
    if i > 1:
        validation += df[df['employer'] == k]['vacancy_id'].sample(frac=valid_part, replace=False, random_state=42).to_list()

validation += df[df['num_samples'] == 1]['vacancy_id'].sample(frac=valid_part, replace=False, random_state=42).to_list()

dill.dump(validation, open('validatin.bin', 'wb'))
