==> Notebooks/05. Length prediction.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

qq('start')

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections
import datetime
from sklearn.preprocessing import LabelEncoder
import xgboost as xgb
from sklearn.metrics import mean_squared_error

df = pd.read_parquet('df_tarin.pq')

df.columns

pd.set_option('display.max_columns', 61)

df.head(2)

def diff_month(d1, d2):
    return (d1.year - d2.year) * 12 + d1.month - d2.month

d1 = datetime.date.today()

df['len_text'] = df['description'].str.len()
df['len_name'] = df['name'].str.len()

df.head(2)

df['num_words'] = df['processed_text'].apply(lambda x: len(x.split(' ')))
df['age'] = df['creation_date'].apply(lambda x: diff_month(d1,x))
df['len_ks'] = df['key_skills'].apply(len)
df['len_html'] = df['html_tags'].apply(len)

drop_cols = ['name', 'description', 'creation_date', 'key_skills', 'specializations', 
             'is_test', 'part', 'groups_of_spec', 'len_groups', 'html_tags', 'clear_text', 'processed_text']
cat_cols = ['work_schedule', 'work_experience', 'currency', 'employer', 'employment']

df.drop(columns=drop_cols, inplace = True)

df.columns

df.head(2)

for col in cat_cols:
    le = LabelEncoder()
    df[col] = le.fit_transform(df[col].to_list())

df.head(2)

model = xgb.XGBRegressor(n_estimators=200, max_depth = 10, objective ='reg:squarederror', learning_rate = 0.1, alpha=10, n_jobs=10)

df['spec_len'] = df['spec_len'].astype(int)

validation = dill.load(open('validatin.bin', 'rb'))

validation = set(validation)

df['is_vailidation'] = df['vacancy_id'].apply(lambda x: x in validation)

df_train = df[~df['is_vailidation']]
df_val = df[df['is_vailidation']]

print(df_train.shape, df_val.shape)

model.fit(df_train.drop(columns=['vacancy_id', 'spec_len', 'is_vailidation']), df_train['spec_len'])

y_pred = model.predict(df_val.drop(columns=['vacancy_id', 'spec_len', 'is_vailidation']))

res = mean_squared_error(df_val['spec_len'], np.round(y_pred, 0))
qq(res)
print(res)

y_pred = model.predict(df_train.drop(columns=['vacancy_id', 'spec_len', 'is_vailidation']))
res = mean_squared_error(df_train['spec_len'], np.round(y_pred, 0))
qq(res)
print(res)

model2 = xgb.XGBClassifier(n_estimators=200, max_depth = 10, objective ='multi:softmax', learning_rate = 0.1, alpha=10, n_jobs=10)

model2.fit(df.drop(columns=['vacancy_id', 'spec_len']), df['spec_len'])

y_pred2 = model2.predict(df.drop(columns=['vacancy_id', 'spec_len']))

mean_squared_error(df['spec_len'], y_pred2)
