==> Notebooks/06. Unpivot train.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

qq('start')

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
import numpy as np
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

#from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections
import fasttext

df = pd.read_parquet('train_text.pq')

df.columns

df.drop(columns=['groups_of_spec', 'key_list', 'name_list'], inplace=True)

df['text'] = df[['employer', 'name', 'key_skills', 'processed_text']].apply(' '.join, axis=1)

df.head(2)

df.drop(columns=['employer', 'name', 'key_skills', 'processed_text'], inplace=True)

def inf_vec(txt):
    return model.infer_vector(txt.split(' '))

df['text_vector'] = pqdm(df['processed_text'], inf_vec, n_jobs=10)

df.to_parquet('df_vector.pq')

def unpivot(obj):
    res = []
    for s in obj.specializations:
        new_obj = obj.to_dict()
        new_obj.pop('specializations')
        new_obj['specialization'] = s
        res.append(new_obj)
    return res

df=pd.read_parquet('df_vector.pq')

df.drop(columns=['processed_text', 'name', 'key_skills', 'groups_of_spec'], inplace = True)

df.head()

df.to_parquet('df_vector.pq')

test_vac_id = set(dill.load(open('validatin.bin','rb')))

df['validation'] = df['vacancy_id'].apply(lambda x: x in test_vac_id)

res = []
for i, r in tqdm(df.iterrows(), total = df.shape[0]):
    res += unpivot(r)

df.head()

df['label'] = df['specializations'].apply(lambda x: ' '.join(['__label__'+str(e) for e in x]))

train = df[~df['is_val']][['vacancy_id', 'label', 'text']].copy()

test = df[df['is_val']][['vacancy_id', 'label', 'text', 'specializations']].copy()

train.to_csv('train.txt', index=False, sep=' ', header=False,)

test.to_csv('test.txt', index=False, sep=' ', header=False,)

valid = set(dill.load(open('validatin.bin', 'rb')))

%%time
model = fasttext.train_supervised('train.txt', dim=300, epoch=10, minCount=2, ws=5, neg=10, verbose=3)

def print_results(N, p, r):
    print("N\t" + str(N))
    print("P@{}\t{:.3f}".format(1, p))
    print("R@{}\t{:.3f}".format(1, r))

print_results(*model.test('train.txt'))

print_results(*model.test('test.txt'))

p = 0.596
r = 0.255

2*p*r/(p+r)

qq('done')

df['is_val'] = df['vacancy_id'].isin(valid)

test = df[df['is_val']]

test = df_unpivoted['is_val']

# test['text'] = test[['employer', 'name', 'key_skills', 'processed_text']].apply(' '.join, axis=1)

# test.drop(columns=['groups_of_spec', 'key_list', 'name_list', 'employer', 'name', 'key_skills', 'processed_text'], inplace=True)

test.head()

test.iloc[0].text

model.predict(test.iloc[0].text, k=6)

test = test.copy()

test.head()

%%time
test['predict'] = test['text'].apply(model.predict, k=6)

test.head()

test[test['vacancy_id']==2700017].predict.values

%%time
test['pred']=test['predict'].apply(lambda x: [int(e.replace('__label__', '').replace('"','')) for e in x[0]])

def f1score(y_true, y_pred):
    y_true = set(y_true)
    y_pred = set(y_pred)
    
    tp = len(y_true & y_pred)
    precision = tp / len(y_pred)
    recall = tp / len(y_true)
    if precision == 0.0 and recall == 0.0:
        score = 0.0
    else:
        score = 2 * precision * recall / (precision + recall)
    return score

def mean_f1score(y_true, y_pred):
    lst = []
    for vacancy_id, true_specs in y_true.items():
        lst.append(f1score(true_specs, y_pred[vacancy_id]))
    return np.mean(lst)

test.head(2)

y_true = test['specializations'].to_dict()

#test['pred'] = test['pred'].apply(lambda x: x[:3])

y_pred = test['pred'].to_dict()

mean_f1score(y_true, y_pred)
