==> Notebooks/09. Prep test - performance test.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

import faiss

qq('start')

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections
import gc

df = pd.read_pickle("df.pkl")

df.columns

df_test = df[df.is_test == True].copy()
del df

gc.collect()

df_test['clear_text'] = df_test.description.apply(remove_html)

qq('cleared')

df_test.to_parquet('df_test.pq')

df_test = pd.read_parquet('df_test.pq')

gc.collect()

%%time
if __name__ ==  '__main__': 
    df_test['processed_text'] = parallelize_on_rows(df_test.clear_text, preprocess_text) 

# %%time
# df_test['processed_text'] = df_test.clear_text.apply(preprocess_text) 

qq('processed')

df_text = df_test[['vacancy_id', 'name', 'specializations', 'key_skills', 'employer', 'groups_of_spec', 'processed_text']].copy()
del df_test
gc.collect()

df_text['key_skills'] = df_text['key_skills'].apply(' '.join)

df_text.to_parquet('test_text.pq')

df_text = pd.read_parquet('test_text.pq')

%%time
if __name__ ==  '__main__':
    df_text['key_skills'] = parallelize_on_rows(df_text.key_skills, preprocess_text)

%%time
if __name__ ==  '__main__':
    df_text['name'] = parallelize_on_rows(df_text.name, preprocess_text)

%%time
df_text['key_list'] = df_text['key_skills'].apply(lambda x: x.split(' '))
df_text['name_list'] = df_text['name'].apply(lambda x: x.split(' '))

df_text.to_parquet('test_text.pq')

%%time
df_text['text_list'] = df_text['processed_text'].apply(lambda x: x.split(' '))

df_text.drop(columns = ['text_list'], inplace = True)

model = Doc2Vec.load('models/vector_size:300_min_count:2_epochs:10_window:7_seed:42_workers:4_negative:50.bin')

def inf_vec(txt):
    return model.infer_vector(txt.split(' '))

df_text['text_vector'] = pqdm(df_text['processed_text'], inf_vec, n_jobs=10)

df_text.head()

df_text.to_parquet('df_vector_test.pq')

res = model.docvecs.most_similar([df_text.loc[0].text_vector], topn=10)

def get_similar(vect):
    return model.docvecs.most_similar([vect], topn=3)

df_text= pd.read_parquet('df_vector_test.pq')

%%time
df_text.iloc[:1000]['text_vector'].apply(get_similar)

df_text.shape[0]/1000*80/3600

%%time
df_text['similar']=df_text['text_vector'].apply(get_similar)

qq('similarity - done')

vac_id = [id for id, sim in res]

df_tarin = pd.read_parquet('df_tarin.pq')

print(df_text.iloc[0].processed_text)

print(df_tarin[df_tarin.vacancy_id == vac_id[1]].clear_text.values)
