==> Notebooks/10. Faiss experiment.ipynb <==
%load_ext autoreload
%autoreload 2
import telepot
bot = telepot.Bot('436447719:AAGq_iK2hE3cPFhmL9Jh53g9HyVkw3SXZbg')
def qq(txt):
    bot.sendMessage(1114926, txt)

import json
import gzip
from itertools import islice
from collections import Counter, defaultdict
from operator import itemgetter
import pyarrow.parquet as pq
import pyarrow as pa

import pandas as pd
from tqdm.notebook import tqdm
import dill
import os
from pqdm.processes import pqdm
import re

from workers import *
    
from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.callbacks import CallbackAny2Vec

import random
import collections
import gc
import faiss

train = pd.read_parquet('df_vector.pq')
test = pd.read_parquet('df_vector_test.pq')

train.head(2)

test.head(2)

train_vectors = dill.load(open('train_vectors.bin', 'rb'))
test_vectors = dill.load(open('test_vectors.bin', 'rb'))
train_specs = dill.load(open('train_spec.bin', 'rb'))

test_vectors = np.array(test_vectors)

print( train_vectors.shape, test_vectors.shape, len(train_specs))

# res = faiss.StandardGpuResources()
# flat_config = faiss.GpuIndexFlatConfig()
# flat_config.device = 0
# faiss.gpu_profiler_start()
# index = faiss.GpuIndexFlatL2(res, 300, flat_config)

index = faiss.IndexFlatL2(400)

index.metric_type = faiss.METRIC_INNER_PRODUCT

%%time
faiss.normalize_L2(train_vectors)
faiss.normalize_L2(test_vectors)

index.add(train_vectors)

print(index.ntotal)

np.savez_compressed('norm_vectors', train=train_vectors, test=test_vectors)

%%time
D, I = index.search(test_vectors[:10000], 10)

1456325/10000*60/60

D[0]

I[0]

from sklearn.metrics.pairwise import cosine_similarity

cosine_similarity([test_vectors[0]], [train_vectors[752562]])

test_indexes = dill.load(open('test_indexes.bin', 'rb'))
train_indexes = dill.load(open('train_indexes.bin', 'rb'))

test_ind = test_indexes[1]
test_ind

train_ind = train_indexes[752562]
train_ind

test[test.vacancy_id == 2700002].processed_text.values[0]

train[train.vacancy_id.isin([train_indexes[id] for id  in I[1]])].processed_text.values

train[train.vacancy_id.isin([train_indexes[id] for id  in I[0]])][['vacancy_id','specializations']]

test_dict = {e:i for i,e in enumerate(test_indexes)}

def get_first_spec(vacancy_id):
    vind = test_dict[vacancy_id]
    return train_specs[int(train_indexes[I[vind][0]])]

def get_all_specs(vacancy_id):
    Is = I[test_dict[vacancy_id]]
    
    e = []
    l = []
    for sim in Is:
        elm = train_specs[int(train_indexes[sim])]
        l.append(len(elm))
        e += list(elm)
    c = Counter(e)
    return [k for k, _ in c.most_common(int(round(np.mean(l),0)))]

vacancy_id = 2700002
from collections import defaultdict
Is = I[test_dict[vacancy_id]]
Ds = D[test_dict[vacancy_id]]
e= []
l = []
for s_ind, sim in enumerate(Is):
    elm = train_specs[int(train_indexes[sim])]
    l.append(len(elm))
    e += [(eo, Ds[s_ind]) for eo in elm]
dd= defaultdict(float)
for spec, wght in e:
    dd[spec] +=  wght
#int(round(np.average(l, weights=Ds),0))
[s for s,w in sorted(dd.items(), key=lambda item: item[1], reverse = True)[:int(round(np.average(l, weights=Ds),0))]]

def get_weigted_specs(vacancy_id):
    Is = I[test_dict[vacancy_id]]
    Ds = D[test_dict[vacancy_id]]
    e= []
    l = []
    for s_ind, sim in enumerate(Is):
        elm = train_specs[int(train_indexes[sim])]
        l.append(len(elm))
        e += [(eo, Ds[s_ind]) for eo in elm]
    dd= defaultdict(float)
    for spec, wght in e:
        dd[spec] +=  wght
    #int(round(np.average(l, weights=Ds),0))
    return [s for s,w in sorted(dd.items(), key=lambda item: item[1], reverse = True)[:int(round(np.average(l, weights=Ds),0))]]

def get_weigted_specs1(vacancy_id):
    Is = I[test_dict[vacancy_id]]
    Ds = D[test_dict[vacancy_id]]
    e= []
    l = []
    for s_ind, sim in enumerate(Is):
        elm = train_specs[int(train_indexes[sim])]
        l.append(len(elm))
        e += [(eo, Ds[s_ind]) for eo in elm]
    dd= defaultdict(float)
    for spec, wght in e:
        dd[spec] +=  wght
    #int(round(np.average(l, weights=Ds),0))
    return [s for s,w in sorted(dd.items(), key=lambda item: item[1], reverse = True)[:5]]

get_weigted_specs(2700002)

get_weigted_specs1(2700002)

get_all_specs(2700002)

get_first_spec(2700002)

data_path = '..\\Data\\In\\'
submissions_path =  '..\\Submissions\\'

sub = []
for vacancy_id in tqdm(test_indexes, total = len(test_indexes)):
    sub.append({'vacancy_id': vacancy_id , 'specializations': get_weigted_specs1(vacancy_id)})

sample_submission = pd.DataFrame(sub)

sample_submission['specializations'] = sample_submission['specializations'].apply(lambda x: x[:3])

# %%time
# sample_submission = pd.DataFrame([
#     (vacancy_id, get_first_spec(vacancy_id))
#     for vacancy_id in test_indexes
# ], columns=['vacancy_id', 'specializations'])
# sample_submission.to_csv(submissions_path+f'sample_submission_{"2020-05-25"}_{1}.csv.gz', index=False, compression='gzip')

sample_submission.head()

sample_submission.loc[0].specializations

sample_submission.to_csv(submissions_path+f'submission_{"2020-07-23"}_{"v5"}.csv.gz', index=False, compression='gzip')
